requirements.txt
# ================================
# Backlog2Notion - Production Ready
# ================================

# === Web Framework ===
fastapi==0.115.13
uvicorn[standard]==0.34.3

# === Data Validation ===
pydantic==2.11.7
pydantic-settings==2.9.1

# === HTTP Client ===
requests==2.32.4

# === Configuration ===
python-dotenv==1.1.0

# ================================
# 依存関係は自動解決されます:
# - starlette (FastAPI互換版)
# - pydantic-core, annotated-types
# - urllib3, certifi, charset-normalizer, idna
# - anyio, sniffio, h11
# - typing-extensions, click
# ================================

# ================================
# Development Dependencies (Optional)
# Uncomment for development:
# ================================
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.12.0
# flake8>=7.0.0
# mypy>=1.8.0
